melhora de front end

+++ 4.py
@@ -1,5 +1,5 @@
 print("A gasolina está R$ 4,30.")
-pgaso=int(4.30)
+pgaso=float(4.30)
 litros=int(input("Quantor o senhor(a) que abastecer em reais: "))
 resultado= litros // pgaso
 if resultado > 1:
